Code cleaning
Code has been refactored in order to improve readability and dispose of obsolete scripts and functions.
