+++ RPGTable/Assets/Scripts/GameManager.cs
@@ -1,11 +1,7 @@
 using Unity.Netcode;
 using UnityEngine;
 
-public class GameManager : MonoBehaviour
+public class GameManager : NetworkBehaviour
 {
 
-    private void OnConnectedToServer()
-    {
-        
-    }
 }
