--- RPGTable/Assets/Scripts/Mover.cs
@@ -1,19 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.AI;
-
-public class Mover : MonoBehaviour
-{
-    NavMeshAgent navMeshAgent;
-
-    private void Start()
-    {
-        navMeshAgent = GetComponent<NavMeshAgent>();
-    }
-
-    public void MoveTo(Vector3 destination)
-    {
-        navMeshAgent.Warp(destination);
-    }
-}
