--- RPGTable/Assets/Scripts/Mover.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9c65d197939bcdd47972c33eb62e3053
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
